requirements.txt
flask==2.3.2
flask-cors==4.0.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
joblib==1.3.2
tensorflow-cpu==2.13.0
gunicorn==21.2.0